particle data rewrite: unit testss
- Switched over C++ unit tests to new particle data arrays
   (tested/working on CPU+GPU)
- corrected some mistakes introduced in the last commits
- fixed bug in ArrayHandle scoping in TwoStepNPH* occuring in debug mode
- still need to test python integration

refs #206

+++ libhoomd/updaters/IntegrationMethodTwoStep.cc
@@ -169,7 +169,7 @@ void IntegrationMethodTwoStep::validateGroup()
     for (unsigned int gidx = 0; gidx < m_group->getNumMembers(); gidx++)
         {
         unsigned int tag = m_group->getMemberTag(gidx);
-        if (m_pdata->getBody(tag) != (int)NO_BODY)
+        if (m_pdata->getBody(tag) != NO_BODY)
             {
             cerr << endl;
             cerr << "***Error! Particle " << tag << " belongs to a rigid body. "
