+++ libhoomd/updaters/TwoStepNVERigid.cc
@@ -616,7 +616,7 @@ void TwoStepNVERigid::validateGroup()
     for (unsigned int gidx = 0; gidx < m_group->getNumMembers(); gidx++)
         {
         unsigned int tag = m_group->getMemberTag(gidx);
-        if (m_pdata->getBody(tag) == (int)NO_BODY)
+        if (m_pdata->getBody(tag) == NO_BODY)
             {
             cerr << endl;
             cerr << "***Error! Particle " << tag << " does not belong to a rigid body. "
